fix round 3: inline NavigationBarItem into NavigationBar scope, onBufferReceived ByteArray, items(size)

+++ app/src/main/java/com/titanali/app/ui/voice/ListenerManager.kt
@@ -27,7 +27,7 @@ class ListenerManager(private val context: Context) {
             override fun onReadyForSpeech(params: Bundle?) {}
             override fun onBeginningOfSpeech() {}
             override fun onRmsChanged(rmsdB: Float) {}
-            override fun onBufferReceived(buffer: ShortArray?) {}
+            override fun onBufferReceived(buffer: ByteArray?) {}
             override fun onEndOfSpeech() {}
 
             override fun onError(error: Int) {
